#271938 Remove `icmaa-tracking`
