--- src/modules/icmaa-tracking/README.md
@@ -1,24 +1,0 @@
-# `icmaa-tracking` module
-
-Get order tracking information from API.
-
-## Configs
-
-```json
-...
-{
-  "icmaa_tracking": {
-    // Endpoint of the api
-    "endpoint": "/api/ext/icmaa-tracking",
-    // The tracking-link is only shown for orders in this status
-    "orderStatusWhitelist": ["complete"],
-    "paqato": {
-      "customerId": XXXXXX
-    }
-  }
-}
-```
-
-## Todo
-
-[ ] ...
