--- src/modules/icmaa-tracking/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "icmaa-tracking",
-  "version": "1.0.0",
-  "author": "cewald <c.ewald@impericon.com> and contributors",
-  "main": "index.js",
-  "license": "MIT",
-  "private": true,
-  "devDependencies": {
-    "@vue-storefront/core": "^1.11.1"
-  },
-  "peerDependencies": {
-    "@vue-storefront/core": "^1.11.1"
-  }
-}
