--- src/modules/icmaa-tracking/routes.ts
@@ -1,5 +1,0 @@
-const Tracking = () => import(/* webpackChunkName: "vsf-icmaa-order-tracking" */ 'icmaa-tracking/pages/Tracking.vue')
-
-export default [
-  { name: 'order-tracking', path: '/order-tracking', component: Tracking }
-]
