--- src/modules/icmaa-tracking/store/mutation-types.ts
@@ -1,2 +1,0 @@
-export const ICMAA_TRACKING_ADD_ORDER = 'ADD_ORDER'
-export const ICMAA_TRACKING_CLR_ORDER = 'CLR_ORDER'
