--- src/modules/icmaa-tracking/types/TrackingState.ts
@@ -1,9 +1,0 @@
-export interface Order {
-  orderId: number,
-  id: string,
-  url: string
-}
-
-export default interface Tracking {
-  orders: Order[]
-}
